move meta-header of the container sub-module to its folder and update the range meta-header
