+++ include/seqan3/range/all.hpp
@@ -42,7 +42,7 @@
 
 #include <seqan3/range/concept.hpp>
 #include <seqan3/range/action/all.hpp>
-#include <seqan3/range/container.hpp>
+#include <seqan3/range/container/all.hpp>
 #include <seqan3/range/view.hpp>
 
 /*!\defgroup range Range
